ciao-launcher: Log docker network creation failures
Any errors reported when creating a docker network are ignored, in case the
error already exists.  It's okay to ignore the error, as the container will
fail to launch if the network does not exist.  However, when this actually
happens it's not obvious why the container failed to launch.  This log
message should help explain things.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>