configuration: Add the ChildUser configuration option
The ChildUser configuration option specifies which user and group launcher should
use to run its sub processes.  It launches a number of sub processes directly,
e.g., qemu and xorriso, which do not need to run with the same privileges as
launcher.  If this value is not set, launcher will launch all subprocesses under
its own user and these processes will share all the same privileges as launcher
itself.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>