ciao-controller: Remove debug installation dependencies
There's no real reason to install sqlite3 in debug builds as it's not used
directly by any of the ciao programs.  Also the existence of deps_debug
is actually breaking ciao on debug builds as it overrides the dependencies
in deps.go.  This means that qemu-img does not get installed, without which
ciao-controller will not work.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>