ciao-deploy: Add unit file for osprepare
This commit adds a new osprepare unit file for each service.  The existing unit
files depend on this new osprepare unit file and will not run until the process
launched by the osprepare unit file has finished.  And the process in the
osprepare unit file is the service run with a --osprepare option.  All the ciao
services have this option.  When it is supplied the services install any
missing dependencies via osprepare and exit.  The osprepare unit file runs the
ciao services as root so they can install their dependencies.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>