ciao-deploy, singlevm: Rename the ciao user to ciaou
Controller, scheduler and processes spawned by launcher now run under the
user ciaou, rather than the user ciao.  The reason for the change is that
the ciao user is predefined in clearlinux and so we get an error when
creating and deleting it.  We could modify ciao deploy to use the ciao
user if it already exists, but then we would need to store some state
somewhere to indicate whether the ciao user had been created by ciao-deploy
on each node and if so that it should be deleted by unjoin and cleanup.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>